test: 拆分性能集成测试结构
